tests.workflows: refactor before reuse
* Split some common functions that will be reused for future
  integration tests.
